# tts: send totalChunks when closing stream