## Summary
- send the total chunk count when sending the final PUT request for a streaming TTS session
- flag end-of-stream flushes so the buffer calls the service with the final-chunk metadata and clears sequence tracking
- extend TextToSpeechService and InputBuffer unit tests to cover totalChunks behavior

## Testing
- npm test

------
https://chatgpt.com/codex/tasks/task_e_68de56b1b0e4832aa9d418be67b99c71